requirements.txt
fastapi
uvicorn
langchain
chromadb
pydantic
python-multipart
loguru
slowapi
pytest
httpx
langchain-community
langchain-openai
langchain-text-splitters
langchain-ollama
pypdf
python-docx